feat: create solar-cli and meta crate solar
