+++ Cargo.toml
@@ -77,6 +77,7 @@ inherits = "profiling"
 [workspace.dependencies]
 # compiler crates
 solar-ast = { version = "0.0.0", path = "crates/ast" }
+solar-cli = { version = "0.0.0", path = "crates/cli" }
 solar-config = { version = "0.0.0", path = "crates/config" }
 solar-data-structures = { version = "0.0.0", path = "crates/data-structures" }
 solar-interface = { version = "0.0.0", path = "crates/interface" }
